+++ app/src/main/java/com/gemini/energy/wordDocGenerator/SorterForWordDocumentGenerator.kt
@@ -707,7 +707,7 @@ class SorterForWordDocumentGenerator {
                             "Griddle",
                             single.energyPowerChange(),
                             costElectricity,
-                            single.age,
+                            single.age.toDouble(),
                             materialCost,
                             implementationCost,
                             savings,
